updated all the routes reports to all events
