+++ routes/api/allEvents.js
@@ -1,0 +1,8 @@
+const express = require("express");
+const router = express.Router();
+const allEventsCtrl = require("../../controllers/api/allEvents");
+
+router.get("/", allEventsCtrl.show);
+router.get("/:id", allEventsCtrl.showOne);
+
+module.exports = router;
